+++ environment.sh
@@ -8,9 +8,12 @@ export FLORA_DB_PASSWORD="postgres"
 export FLORA_DB_DATABASE="flora_dev"
 export FLORA_DB_POOL_CONNECTIONS="100"
 export FLORA_DB_TIMEOUT="10"
+export FLORA_DB_SSLMODE="allow"
+export FLORA_DB_PARAMETERS="?sslmode=verify-ca"
 
-export FLORA_PG_URI="postgresql://${FLORA_DB_USER}:${FLORA_DB_PASSWORD}@${FLORA_DB_HOST}:${FLORA_DB_PORT}/${FLORA_DB_DATABASE}"
-export FLORA_PG_CONNSTRING="host=${FLORA_DB_HOST} dbname=${FLORA_DB_DATABASE} user=${FLORA_DB_USER} password=${FLORA_DB_PASSWORD}"
+export FLORA_DB_CONNSTRING="host=${FLORA_DB_HOST} dbname=${FLORA_DB_DATABASE} \
+  user=${FLORA_DB_USER} password=${FLORA_DB_PASSWORD} \
+  sslmode=${FLORA_DB_SSLMODE:-allow}"
 
 export FLORA_HTTP_PORT=8083
 export FLORA_ENVIRONMENT="development"