+++ environment.test.sh
@@ -6,4 +6,6 @@ export FLORA_DOMAIN="localhost"
 
 export FLORA_DB_DATABASE="flora_test"
 
-export FLORA_PG_CONNSTRING="host=${FLORA_DB_HOST} dbname=${FLORA_DB_DATABASE} user=${FLORA_DB_USER} password=${FLORA_DB_PASSWORD}"
+export FLORA_DB_CONNSTRING="host=${FLORA_DB_HOST} dbname=${FLORA_DB_DATABASE} \
+  user=${FLORA_DB_USER} password=${FLORA_DB_PASSWORD} \
+  sslmode=allow"