+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Remove padding from README code paragraphs (#203)
 * Simplify the database pool settings (#210)
 * Readjust the size of the main title across browser sizes (#211)
+* Take the database settings as a libpq connstring (#213)
 
 ## v1.0.2 -- 2022-09-13